+++ src/gfn/containers/__init__.py
@@ -1,4 +1,2 @@
-from .states import States
 from .trajectories import Trajectories
 from .transitions import Transitions
-from .actions import Actions
