+++ src/gfn/containers/transitions.py
@@ -6,9 +6,9 @@ import torch
 from torchtyping import TensorType
 
 if TYPE_CHECKING:
+    from gfn.actions import Actions
     from gfn.envs import Env
-    from gfn.containers.states import States
-    from gfn.containers.actions import Actions
+    from gfn.states import States
 
 from gfn.containers.base import Container
