+++ src/gfn/distributions.py
@@ -5,8 +5,9 @@ from typing import Optional
 import torch
 from torchtyping import TensorType
 
-from gfn.containers import States, Trajectories
+from gfn.containers import Trajectories
 from gfn.envs import Env
+from gfn.states import States
 
 # Typing
 TensorPmf = TensorType["n_states", float]
