+++ src/gfn/samplers/trajectories_sampler.py
@@ -3,9 +3,10 @@ from typing import List, Optional
 import torch
 from torchtyping import TensorType
 
-from gfn.containers import States, Trajectories
+from gfn.containers import Trajectories
 from gfn.envs import Env
 from gfn.samplers import ActionsSampler, BackwardActionsSampler
+from gfn.states import States
 
 # Typing
 StatesTensor = TensorType["n_trajectories", "state_shape", torch.float]
